+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [JavaScript] Add `ProgressPrinter` ([#40](https://github.com/cucumber/pretty-formatter/pull/40))
 - Print execution duration in summary ([#62](https://github.com/cucumber/pretty-formatter/pull/62))
 - Include responsible step in summary ([#63](https://github.com/cucumber/pretty-formatter/pull/63))
+- Show matching step definitions for ambiguous steps ([#67](https://github.com/cucumber/pretty-formatter/pull/67))
 
 ### Changed
 - [JavaScript] BREAKING CHANGE: Expose `PrettyPrinter` rather than a high-level formatter ([#40](https://github.com/cucumber/pretty-formatter/pull/40))